// packages/core/colors.d.ts
export interface ColorsInterface {
  grey: string
  grey50: string
  grey100: string
  grey200: string
  grey300: string
  grey400: string
  grey500: string
  grey600: string
  grey700: string
  grey800: string
  grey900: string
  blue: string
  blue100: string
  blue200: string
  blue500: string
  blue700: string
  red: string
  red100: string
  red200: string
  red500: string
  red700: string
}
export declare const colors: ColorsInterface